Feature: minor optimisations with references to map tile size.

+++ modules/server/src/main/java/com/limpygnome/projectsandbox/server/entity/ai/pathfinding/astar/AStarPath.java
@@ -50,7 +50,7 @@ public class AStarPath extends Path
         }
 
         // Node separation is tile size
-        nodeSeparation = (float) map.tileSize;
+        nodeSeparation = map.tileSize;
 
         // Destroy data structures used during search
         closedNodes.clear();
